* Try three to remove compiler warning while not breaking PRACK support...    (thanks to mikma for protecting the feature from me)
git-svn-id: svn://svn.minisip.org/minisip/trunk@2755 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libmsip/source/dialogs/SipDialog.cxx
@@ -66,7 +66,7 @@ SipDialog::SipDialog(MRef<SipStack*> stack, MRef<SipDialogConfig*> callconf):
 					
 	dialogState.isEarly=false;	//same as for "secure"?! -EE
 	dialogState.isEstablished = false;
-	dialogState.rseqNo = -1;
+	dialogState.rseqNo = (uint32_t)-1;
 }
 
 SipDialog::~SipDialog(){
